small-website-crawler-azure/requirements.txt
# Do not include azure-functions-worker in this file
# The Python Worker is managed by the Azure Functions platform
# Manually managing azure-functions-worker may cause unexpected issues

azure-functions
anyio==4.6.0
apify_client==1.8.1
apify_shared==1.1.2
azure-core==1.31.0
azure-storage-blob==12.23.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
cryptography==43.0.1
exceptiongroup==1.2.2
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
idna==3.10
isodate==0.7.0
more-itertools==10.5.0
pycparser==2.22
requests==2.32.3
six==1.16.0
sniffio==1.3.1
typing_extensions==4.12.2
urllib3==2.2.3